Move import test to resource test file
