--- datadog/tests/import_datadog_api_key_test.go
@@ -1,32 +1,0 @@
-package test
-
-import (
-	"context"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-)
-
-func TestDatadogApiKey_import(t *testing.T) {
-	t.Parallel()
-	resourceName := "datadog_api_key.foo"
-	ctx, accProviders := testAccProviders(context.Background(), t)
-	apiKeyName := uniqueEntityName(ctx, t)
-	accProvider := testAccProvider(t, accProviders)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ProviderFactories: accProviders,
-		CheckDestroy:      testAccCheckDatadogApiKeyDestroy(accProvider),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckDatadogApiKeyConfigRequired(apiKeyName),
-			},
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
