+++ datadog/tests/provider_test.go
@@ -65,13 +65,10 @@ var testFiles2EndpointTags = map[string]string{
 	"tests/data_source_datadog_service_level_objectives_test":          "service-level-objectives",
 	"tests/data_source_datadog_synthetics_locations_test":              "synthetics",
 	"tests/data_source_datadog_synthetics_global_variable_test":        "synthetics",
-	"tests/import_datadog_api_key_test":                                "api_keys",
-	"tests/import_datadog_application_key_test":                        "application_keys",
 	"tests/import_datadog_downtime_test":                               "downtimes",
 	"tests/import_datadog_integration_pagerduty_test":                  "integration-pagerduty",
 	"tests/import_datadog_logs_pipeline_test":                          "logs-pipelines",
 	"tests/import_datadog_monitor_test":                                "monitors",
-	"tests/import_datadog_organization_settings_test":                  "organization",
 	"tests/import_datadog_user_test":                                   "users",
 	"tests/provider_test":                                              "terraform",
 	"tests/resource_datadog_api_key_test":                              "api_keys",
